requirements.txt
alabaster==0.7.12
aniso8601==9.0.1
attrs==21.2.0
awscli==1.19.67
Babel==2.9.1
botocore==1.20.67
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.3
coverage==5.5
cycler==0.10.0
decorator==4.4.2
docutils==0.15.2
flake8==3.9.1
Flask==1.1.2
flask-restx==0.3.0
gunicorn==20.1.0
ibm-cloud-sdk-core==3.9.0
ibm-cos-sdk==2.10.0
ibm-cos-sdk-core==2.10.0
ibm-cos-sdk-s3transfer==2.10.0
ibmcloudant==0.0.33
idna==2.10
imageio==2.9.0
imagesize==1.2.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
jproperties==2.1.0
jsonschema==3.2.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.1
mccabe==0.6.1
networkx==2.5.1
numpy==1.20.2
packaging==20.9
Pillow==8.2.0
pyasn1==0.4.8
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.9.0
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
s3transfer==0.4.2
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
sklearn==0.0
skorch==0.10.0
snowballstemmer==2.1.0
Sphinx==3.5.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
tabulate==0.8.9
threadpoolctl==2.1.0
tifffile==2021.4.8
torch==1.8.1
torchvision==0.9.1
tqdm==4.60.0
typing-extensions==3.10.0.0
urllib3==1.26.4
websocket-client==0.58.0
Werkzeug==1.0.1
